Update request-level validation report (#78)
Account for cases where request has 'technically' invalid samples but the
samples have errors that are passable and can allow them to be imported
with errors.

The request validation report was reporting as having critically failed samples
but the 'samples' reported at the request-level validation report was actually empty.

The request-level validation report should only report 'samples' as an error
when it has samples with critical validation errors like missing cmo patient ids.

Signed-off-by: Angelica Ochoa <15623749+ao508@users.noreply.github.com>